Fix POST operation to post data to data structures
Fix POST to send data
Fix tests to pass travis ci
Update requirements.txt
