+++ .gitignore
@@ -1,4 +1,4 @@
-venv36
+venv3
 UI
 .pytest_cache
 .vscode
