--- data.py
@@ -1,43 +1,0 @@
-users =[
-    {
-        'fname':'', 
-        'lname':'', 
-        'email':'',
-        'password':''
-    }
-]
-
-
-rides = [
-    {   'id' : 1 ,
-        'driver' : 'John Doe' ,
-        'start_loc' : 'Nairobi' , 
-        'end_loc' : 'Thika',
-        'departure_time' : '1800HRS' , 
-        'date' : '13/6/2018' , 
-        'route' : 'Thika Super Highway' , 
-        'cost' : '400'
-    } ,
-    {   
-        'id' : 2 ,
-        'driver' : 'Jane Doe' , 
-        'start_loc' : 'Nairobi', 
-        'end_loc' : 'Syokimau',
-        'departure_time' : '0900HRS', 
-        'date' : '14/6/2018' , 
-        'route' : 'Mombasa Road' , 
-        'cost' : '200'
-    }
-]
-requested = [
-    {
-        'pickup_loc': '',
-    }
-]
-
-dummy_user = {
-        'fname':'John', 
-        'lname':'Doe', 
-        'email':'mail@gmail.com',
-        'password':'pass123'
-    }
